refactor: pwk to key info for file name
