+++ examples/svelte-app/src/components/settings/ExportKeyInfoComponent.svelte
